add testsuite_helpers
